requirements.txt
langgraph>=0.0.18
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0
chromadb>=0.4.22
pydantic>=2.5.0
fastapi>=0.109.0
uvicorn>=0.27.0 